Add player health logic.
Added PlayerHealth script.
Added PlayerUI scripts.
Added GameUI canvas.

+++ Assets/ProjectResources/Characters/Scripts/Interfaces/IDamager.cs
@@ -1,5 +1,5 @@
 ﻿public interface IDamager
 {
-    float Damage { get; set; }
+    int Damage { get; set; }
     void TakeDamage(IHealth target);
 }
